[sim] per-op trace instrumentation for cycle counting
Emit per-op work-counts in the simulator trace so downstream tools can reason about compute vs. movement volume per kernel:

- compute_op event (op_type, tiles, dtype) at each math-op site (matmul, elementwise binary/unary, reduce)
- copy_end enriched with per-locality tile counts + dtype
- ttl.block.fill(0, ...) derives dtype from the target block instead of hardcoding bf16 (unblocks fp32 kernels); example set updated

Additive to the trace schema; no change to existing behavior.

+++ examples/matmul_1d.py
@@ -68,7 +68,9 @@ def matmul_1d(
                     nb = local_nb + node_index * nb_per_node
                     if nb < Nb:
                         with out_dfb.reserve() as out_blk:
-                            acc = ttl.block.fill(0, shape=out_blk.shape)
+                            acc = ttl.block.fill(
+                                0, shape=out_blk.shape, dtype=out_blk.dtype
+                            )
                             for kb in range(Kb):
                                 with a_dfb.wait() as a_blk, b_dfb.wait() as b_blk:
                                     acc += a_blk @ b_blk
